Opravný commit

+++ DeliveryService/src/main/java/pa165/deliveryservice/daoImplementation/PostmanDaoImpl.java
@@ -5,12 +5,20 @@ import pa165.deliveryservice.entity.Postman;
 import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
 
 public class PostmanDaoImpl implements PostmanDao {
 
     private EntityManagerFactory emf;
 
+    public PostmanDaoImpl() {
+        emf = Persistence.createEntityManagerFactory("myUnit");
+    }
+
     public PostmanDaoImpl(EntityManagerFactory emf) {
+        if (emf == null) {
+            throw new NullPointerException();
+        }
         this.emf = emf;
     }
 
@@ -24,26 +32,38 @@ public class PostmanDaoImpl implements PostmanDao {
 
     @Override
     public void updatePostman(Postman postman) {
+        if (postman == null) {
+            throw new NullPointerException("Postman is null.");
+        }
         EntityManager em = emf.createEntityManager();
-        Postman postmandb = em.find(Postman.class, postman.getId());
-        postmandb.setFirstName(postman.getFirstName());
-        postmandb.setLastName(postman.getLastName());
-        postmandb.setDeliveries(postman.getDeliveries());
+        em.getTransaction().begin();
+        em.merge(postman);
+        em.getTransaction().commit();
         em.close();
     }
 
     @Override
     public void deletePostman(Postman postman) {
+        if (postman == null) {
+            throw new NullPointerException("Postman is null.");
+        }
         EntityManager em = emf.createEntityManager();
+        em.getTransaction().begin();
         Postman postmandb = em.find(Postman.class, postman.getId());
         em.remove(postmandb);
+        em.getTransaction().commit();
         em.close();
     }
 
     @Override
     public void addPostman(Postman postman) {
+        if (postman == null) {
+            throw new NullPointerException("Postman is null.");
+        }
         EntityManager em = emf.createEntityManager();
+        em.getTransaction().begin();
         em.persist(postman);
+        em.getTransaction().commit();
         em.close();
     }
